requirements.txt
flask
transformers
torch
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv